requirements.txt
astroid==2.13.5
build==1.2.2.post1
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
databricks-connect==13.3.2
databricks-sdk==0.44.0
dill==0.3.9
google-auth==2.38.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
idna==3.10
isort==5.13.2
lazy-object-proxy==1.10.0
mccabe==0.7.0
mypy==1.15.0
mypy-extensions==1.0.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
protobuf==5.29.3
py4j==0.10.9.7
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pylint==2.15.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
rsa==4.9
six==1.17.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wrapt==1.17.2
